Simplify TypeOption and use RelevantIcons to display them
