+++ content/webapp/components/ExhibitionGuideLinks/TypeOption.tsx
@@ -1,13 +1,17 @@
-import { FunctionComponent, SyntheticEvent } from 'react';
+import { setCookie } from 'cookies-next';
+import { FunctionComponent } from 'react';
+import styled from 'styled-components';
+import cookies from '@weco/common/data/cookies';
 import { IconSvg } from '@weco/common/icons/types';
 import { font } from '@weco/common/utils/classnames';
 import Space from '@weco/common/views/components/styled/Space';
-import styled from 'styled-components';
 import { PaletteColor } from '@weco/common/views/themes/config';
 import Icon from '@weco/common/views/components/Icon/Icon';
 import { plainListStyles } from '@weco/common/views/components/styled/PlainList';
 import { useToggles } from '@weco/common/server-data/Context';
-import { arrow, speechToText } from '@weco/common/icons';
+import { arrow } from '@weco/common/icons';
+import { RelevantIcons } from '@weco/content/pages/guides/exhibitions/[id]/[type]';
+import { ExhibitionGuideType } from '@weco/content/types/exhibition-guides';
 
 export const TypeList = styled(Space).attrs({
   $v: { size: 'l', properties: ['row-gap'] },
@@ -49,6 +53,18 @@ const TypeLink = styled.a<{
   }
 `;
 
+function cookieHandler(key: string, data: string) {
+  // We set the cookie to expire in 8 hours (the maximum length of
+  // time the galleries are open in a day)
+  const options = {
+    maxAge: 8 * 60 * 60,
+    path: '/',
+    secure: true,
+  };
+  setCookie(key, data, options);
+}
+
+// TODO Review how this can be streamlined when we move to the new EG models
 type Props = {
   url: string;
   title: string;
@@ -59,10 +75,8 @@ type Props = {
     | 'accent.lightGreen'
     | 'accent.lightPurple'
     | 'accent.lightBlue';
-
+  type: ExhibitionGuideType;
   icon?: IconSvg;
-  hasTranscripts?: boolean;
-  onClick?: (event: SyntheticEvent<HTMLAnchorElement>) => void;
 };
 
 const TypeOption: FunctionComponent<Props> = ({
@@ -71,10 +85,13 @@ const TypeOption: FunctionComponent<Props> = ({
   text,
   backgroundColor,
   icon,
-  hasTranscripts,
-  onClick,
+  type,
 }) => {
   const { egWork } = useToggles();
+
+  const onClick = () => {
+    cookieHandler(cookies.exhibitionGuideType, type as string);
+  };
 
   return egWork ? (
     <TypeItem $egWork={egWork}>
@@ -89,20 +106,9 @@ const TypeOption: FunctionComponent<Props> = ({
           $h={{ size: 'm', properties: ['padding-left', 'padding-right'] }}
         >
           <h2 className={font('wb', 3)}>{title}</h2>
-          {icon && (
-            <Icon icon={icon} sizeOverride="height: 32px; width: 32px;" />
-          )}
-          {hasTranscripts && (
-            <Space
-              $h={{ size: 's', properties: ['margin-left'] }}
-              style={{ display: 'inline' }}
-            >
-              <Icon
-                icon={speechToText}
-                sizeOverride="height: 32px; width: 32px;"
-              />
-            </Space>
-          )}
+
+          <RelevantIcons types={[type]} />
+
           <div style={{ position: 'absolute', bottom: '10px', right: '15px' }}>
             <Icon icon={arrow} sizeOverride="height: 32px; width: 32px;" />
           </div>
